requirements.txt
# Core dependencies
apache-airflow[postgres,s3]==2.8.1
polars==0.20.3
boto3==1.34.0
botocore==1.34.0

# API and HTTP clients
requests==2.31.0
aiohttp==3.9.1

# Data processing
pandas==2.1.4
numpy==1.24.3
pydantic==2.5.2

# Database
# psycopg2-binary==2.9.9
# sqlalchemy==1.4.46
sqlalchemy==1.4.51
psycopg2-binary==2.9.10
# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# AWS SDK
awscli==1.32.0

# Tableau connectivity (optional)
tableauserverclient==0.28 